feat: add new mode to only add when missing
This change will instruct Neutron to only add records when missing from the DB.
This is crucial when running a cluster that leverages the same OVN deployment
outside of OpenStack. The changes will ensure that the sync tool doesn't delete
records in any way which will ensure there's no accidental record loss due to
item prefix or labels being missing.

Change-Id: Ib38c9ad22ecb7f59659e7947782b20a432433a76
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>